requirements.txt
psycopg2
psycopg2-binary
streamsync
pandas
python-dotenv